requirements.txt
absl-py==1.4.0
appdirs
astunparse==1.6.3
blobfile==2.0.0
brotlipy
cachetools==5.3.0
certifi==2022.12.7
cffi
charset-normalizer==3.0.1
click
contourpy==1.0.6
cryptography
cycler==0.11.0
docker-pycreds==0.4.0
einops==0.6.0
filelock==3.8.2
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
gitdb
GitPython
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna
importlib-metadata==6.0.0
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.2
mpi4py
numpy==1.24.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pathtools==0.1.2
Pillow==9.3.0
protobuf==3.20.1
psutil
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
pycryptodomex==3.16.0
pyOpenSSL
pyparsing==3.0.9
PySocks
python-dateutil==2.8.2
pytz==2022.7
PyYAML
requests
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.9.3
sentry-sdk
setproctitle
six
smmap
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions
urllib3
wandb
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
